Ensure the "longrepr" attribute is set on failing tests
Some pytest functionality relies of the fact that the "longrepr"
attribute is not None for failing tests. According to pytest type
annotations this can be a string explaining why the test failed or some
complicated traceback object. As we cannot easily make a traceback
object, we can simply put a string explaining why the test failed.