Change longrepr report to include memory info
Refactor limit_memory to return a schema-based type, which
can be converted to a PytestSectio and a longrepr string.

Remove item.nodeid from longrepr since longrepr is only used by
junit-xml parser and the structure of the xml document
makes the item.nodeid in the error message redundant.

Add a test that parses the resulting xml file to make sure
the expected string was written.