Fix a flaky test
When this test was selected to be run with `pytest -k` or `--lf`, it
would fail because of allocations made by the platform threading
implementation:

    - 8.0MiB allocated here:
        __pthread_create_2_1
        PyThread_start_new_thread
        ...

When it was run as part of the larger test suite, it would pass, because
the test suite had already started other threads, and paid that cost in
a place where we weren't looking for leaks.

Since we can't tell pthreads not to cache the pthread_t for later reuse,
the best we can do here is just raise all of our thresholds for this
test by a constant factor, so that the ones that are supposed to fail
will fail with or without an extra 8 MB allocation, and the ones that
are supposed to pass will pass with or without it as well.

This commit scales all of the sizes and limits used by this test up by
a factor of 4096 (and drops one of the test vectors, since we don't need
to leak 400 MB just to prove a point when we already know we fail after
leaking 40 MB).

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>