Ensure our histogram uses a non-zero step
When all allocations are the same size, we were falling back to the size
of the smallest allocation as our step for generating histogram buckets.
When the smallest allocation was zero bytes, that would lead to
a division by zero and an exception.

Fix this by falling back to 1 instead of the size of the smallest
allocation. This will never lead to a zero division error, and still
puts every element in the same bucket because they're all equal
(otherwise we wouldn't have needed the fallback).

Signed-off-by: Matt Wozniski <mwozniski@bloomberg.net>